Перевірка роботи
